Avoid image deletion if not changed ==> state ifImageChanged: boolean Text area configured for solution.details

+++ src/components/SolutionForm.tsx
@@ -1,5 +1,4 @@
 import React, { useState, useEffect } from "react";
-import ReactMarkdown from "react-markdown";
 import { Region, Solution as SolutionType } from "../types";
 import { useNavigate } from "react-router-dom";
 import { updateSolution } from "../reducers/solutionsReducer";
@@ -8,19 +7,37 @@ import { useAuth0 } from "@auth0/auth0-react";
 import SolutionContactForm from "./SolutionContactForm";
 import ImageUploader from "./ImageUploader";
 
+function countOccurrences(string: string, substring: string) {
+	// Escape special characters in the substring
+	const escapedSubstring = substring.replace(/[-/\\^$*+?.()|[\]{}]/g, "\\$&");
+
+	// Create a regular expression to match the substring globally
+	const regex = new RegExp(escapedSubstring, "g");
+
+	// Use match() to get an array of all occurrences of the substring
+	const matches = string.match(regex);
+
+	// Return the number of occurrences
+	return matches ? matches.length : 0;
+}
+
 interface Props {
 	solution: SolutionType;
 }
 
 const SolutionForm: React.FC<Props> = ({ solution }) => {
+	const solutionDetailsPlaceHolder =
+		"Utilisez la syntaxe MarkDown pour mettre en forme votre texte\nSautez une ligne pour séparer les paragraphes\n\n#Titre\nVotre texte de description\n\n##Sous-titre\nExemple de liste :\n* Item 1\n* Item 2\n* Item 3\n\n##Exemple de lien url :\n[Texte de mon lien](https://monlien.fr)";
 	const [name, setName] = useState(solution.name);
 	const [description, setDescription] = useState(solution.description);
 	const [region, setRegion] = useState(solution.region);
 	const [image, setImage] = useState(solution.img);
 	const [imageId, setImageId] = useState(solution.imgId);
-	//const [googlePlusCode, setGooglePlusCode] = useState(solution.googlePlusCode);
+	const [imageModified, setImageModified] = useState(false);
+	const [googlePlusCode, setGooglePlusCode] = useState(solution.googlePlusCode);
 	const [contact, setContact] = useState(solution.contact);
-	//const [website, setWebsite] = useState(solution.website);
+	const [website, setWebsite] = useState(solution.website);
+	const [details, setDetails] = useState(solution.details);
 	const [authorizedUser, setAuthorizedUser] = useState(true);
 	const navigate = useNavigate();
 	const dispatch = useAppDispatch();
@@ -34,6 +51,8 @@ const SolutionForm: React.FC<Props> = ({ solution }) => {
 			return null;
 		}
 	};
+
+	console.log(solution.details);
 
 	useEffect(() => {
 		const checkScope = async () => {
@@ -74,13 +93,30 @@ const SolutionForm: React.FC<Props> = ({ solution }) => {
 		setRegion(e.target.value as Region);
 	};
 
-	const handleImageUpload = (image: { url: string; id: string }) => {
+	const handleImageUpload = (image: {
+		url: string;
+		id: string;
+		modified: boolean;
+	}) => {
 		setImage(image.url);
 		setImageId(image.id);
+		setImageModified(image.modified);
 	};
 
 	const handleContactChange = (contact: string) => {
 		setContact(contact);
+	};
+
+	const handleWebsiteChange = (website: string) => {
+		setWebsite(website);
+	};
+
+	const handleGooglePlusCodeChange = (googlePlusCode: string) => {
+		setGooglePlusCode(googlePlusCode);
+	};
+
+	const handleDetailsChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+		setDetails(e.target.value);
 	};
 
 	const handleLogin = async (e: React.MouseEvent<HTMLButtonElement>) => {
@@ -118,9 +154,10 @@ const SolutionForm: React.FC<Props> = ({ solution }) => {
 						region: region,
 						img: image,
 						imgId: imageId,
-						//googlePlusCode: googlePlusCode,
+						googlePlusCode: googlePlusCode,
 						contact: contact !== "" ? contact : solution.contact,
-						//website: website,
+						website: website !== "" ? website : solution.website,
+						details: details !== "" ? details : solution.details,
 					},
 					accessToken
 				)
@@ -130,14 +167,15 @@ const SolutionForm: React.FC<Props> = ({ solution }) => {
 			//////////////////////////////
 			// Add a condition here ! Do not delete default image ! //
 			///////////////////////////////
-
-			await fetch(`http://localhost:3000/imagekit/${solution.imgId}`, {
-				method: "DELETE",
-				headers: {
-					Authorization: "Bearer " + accessToken,
-					"Content-Type": "application/json",
-				},
-			});
+			if (imageModified) {
+				await fetch(`/api/imagekit/${solution.imgId}`, {
+					method: "DELETE",
+					headers: {
+						Authorization: "Bearer " + accessToken,
+						"Content-Type": "application/json",
+					},
+				});
+			}
 		} catch (error) {
 			console.log(error);
 		}
@@ -145,6 +183,20 @@ const SolutionForm: React.FC<Props> = ({ solution }) => {
 
 	const handleCancelChange = () => {
 		navigate(`/solutions/${solution.id}`);
+	};
+
+	const newlineCount = solution.details
+		? countOccurrences(solution.details, "\n")
+		: countOccurrences(solutionDetailsPlaceHolder, "\n");
+	const textareaStyle = {
+		height: `${newlineCount * 17}px`,
+	};
+
+	const handleInput: React.FormEventHandler<HTMLTextAreaElement> = (event) => {
+		const textarea = event.target as HTMLTextAreaElement;
+		textarea.style.height = "";
+		textarea.style.height = `${textarea.scrollHeight}px`;
+		window.scrollTo(0, document.body.scrollHeight);
 	};
 
 	return (
@@ -238,13 +290,20 @@ const SolutionForm: React.FC<Props> = ({ solution }) => {
 					<SolutionContactForm
 						solution={solution}
 						onContactChange={handleContactChange}
+						onWebsiteChange={handleWebsiteChange}
+						onGooglePlusCodeChange={handleGooglePlusCodeChange}
 					/>
 				</div>
-				{solution.details ? (
-					<ReactMarkdown className="solutionForm__details">
-						{solution.details}
-					</ReactMarkdown>
-				) : null}
+				<div>
+					<textarea
+						className="solutionForm__details"
+						value={details}
+						onChange={handleDetailsChange}
+						placeholder={solutionDetailsPlaceHolder}
+						onInput={handleInput}
+						style={textareaStyle}
+					/>
+				</div>
 			</div>
 		</div>
 	);
